REFACTOR: put things in subdirectories
