+++ source/services/Service_Management.py
@@ -1,10 +1,7 @@
-from google.appengine.api import users
-from google.appengine.ext import ndb
 import webapp2
-import re
-import json
-from NdbClasses import *
-from Service_Utils import *
+
+from source.models.NdbClasses import *
+from source.services.Service_Utils import *
 
 
 # stream management service
